requirements.txt
fastapi
uvicorn[standard]
requests
pydantic
chromadb
pandas
faiss-cpu
langchain
langchain-ollama
streamlit
SQLAlchemy>=2.0
alembic>=1.13
PyMySQL>=1.1
pydantic-settings